SREP-923: Relocate EBS IOPS reporting exporter (#1)
* SREP-963: Move to use AWS cred minter

As mentioned in the comment in #1, we have a restriction from passing
AWS credentials (and others, I imagine) at deploy time and should look
to the [AWS credential
minter](https://github.com/openshift/cloud-credential-operator) instead.
To that end, I've created https://github.com/lisa/docker-yq-kubectl
which is intended for use as an init container in support of the
credential minter operator. Usage of the init container image is
documented in its repository, but in summary:

The basic cluster installation (as OSD) provides setup information in
the `kube-system` namespace as `configmaps/cluster-config-v1`, including
the AWS region into which the cluster is installed. That is an important
piece of information for the credentials piece because the cloud
credential operator does not include the region information. In order to
access that at runtime, the service account this exporter runs as
requires a permissions to access that configmap object; it must also be
accessed from inside the pod with `kubectl`, as there is no support in
Kubernetes to directly reference resources in a different namespace.

This is based on feedback for one of these SREP-922/923/963 PRs: A
desire to have `sre-` prefixed to objects created for and by SRE. I've
taken it one step further to standardize on the names of objects, driven
by the Makefile.

Signed-off-by: Lisa Seelye <lseelye@redhat.com>